refactor(authz): extract PolicyResult to types.py and remove wrapper functions
Breaks the circular import between dependencies.py and policies/ by moving
PolicyResult and PolicyFn to a dedicated types module. This allows guards
to reference policy functions directly (e.g. finance.can_read) instead of
through intermediary _finance_can_read() wrappers.
