+++ server/polar/authz/policies/finance.py
@@ -1,6 +1,6 @@
 from polar.account.service import account as account_service
 from polar.auth.models import AuthSubject, Organization, User, is_organization, is_user
-from polar.authz.dependencies import PolicyResult
+from polar.authz.types import PolicyResult
 from polar.models import Organization as OrganizationModel
 from polar.postgres import AsyncReadSession
